+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/dispenser/ThrowableSlimeballBehavior.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.supplementaries.common.block.dispenser;
 
+import net.mehvahdjukaar.supplementaries.common.block.fire_behaviors.ProjectileStats;
 import net.mehvahdjukaar.supplementaries.common.entities.SlimeBallEntity;
 import net.mehvahdjukaar.supplementaries.reg.ModSounds;
 import net.minecraft.core.BlockSource;
@@ -25,16 +26,14 @@ class ThrowableSlimeballBehavior extends ProjectileBehavior {
         return new SlimeBallEntity(source.getLevel(), position.x(), position.y(), position.z());
     }
 
-    //TODO: standardize these
     @Override
     protected float getProjectileInaccuracy() {
-        return 7.0F;
+        return ProjectileStats.SLIMEBALL_DISPENSER_INACCURACY;
     }
 
-    //TODO: fix throwable bricks rendering glitchyness
     @Override
     protected float getProjectileVelocity() {
-        return 0.8F;
+        return ProjectileStats.SLIMEBALL_DISPENSER_SPEED;
     }
 
 }