+++ common/src/main/java/net/mehvahdjukaar/supplementaries/reg/ModFluids.java
@@ -1,6 +1,9 @@
 package net.mehvahdjukaar.supplementaries.reg;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.mehvahdjukaar.moonlight.api.fluids.SoftFluid;
+import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidRegistry;
+import net.mehvahdjukaar.moonlight.api.misc.DynamicHolder;
 import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.mehvahdjukaar.supplementaries.Supplementaries;
 import net.mehvahdjukaar.supplementaries.common.fluids.FiniteFluid;
@@ -25,6 +28,9 @@ public class ModFluids {
     public static final Supplier<FlammableLiquidBlock> LUMISENE_BLOCK;
     public static final Supplier<BucketItem> LUMISENE_BUCKET;
     public static final Supplier<Item> LUMISENE_BOTTLE;
+
+    public static final DynamicHolder<SoftFluid> LUMISENE_SOFT_FLUID = DynamicHolder.of(
+           Supplementaries.res(ModConstants.LUMISENE_NAME), SoftFluidRegistry.KEY);
 
     static{